Structure Refactoring: Phase 05
