+++ lib/shared/services/core/global_pools.dart
@@ -1,6 +1,6 @@
 import 'package:pool/pool.dart';
 
-import 'concurrency_manager.dart';
+import '../../infraestructure/concurrency_manager.dart';
 
 /// Central registry of shared Pool instances.
 ///
