+++ lib/shared/services/media/content_grouping_service.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 
-import '../../concurrency_manager.dart';
+import '../../infraestructure/concurrency_manager.dart';
 import '../../entities/media_entity.dart';
 import '../core/logging_service.dart';
