+++ lib/shared/services/media/duplicate_detection_service.dart
@@ -1,4 +1,4 @@
-import '../../concurrency_manager.dart';
+import '../../infraestructure/concurrency_manager.dart';
 import '../../entities/media_entity.dart';
 import '../core/logging_service.dart';
 import 'media_hash_service.dart';
