+++ lib/steps/step_04_extract_dates/services/date_extractors/json_date_extractor.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 import 'dart:io';
 import 'package:coordinate_converter/coordinate_converter.dart';
-import '../json_metadata_matcher_service.dart';
+import '../../../../shared/services/json_metadata/json_metadata_matcher_service.dart';
 
 /// Finds corresponding json file with info from media file and gets 'photoTakenTime' from it
 Future<DateTime?> jsonDateTimeExtractor(
