+++ lib/steps/step_04_extract_dates/step_04_extract_dates.dart
@@ -1,7 +1,7 @@
 import 'package:console_bars/console_bars.dart';
 
 import '../../shared/entities/media_entity.dart';
-import '../../shared/models/pipeline_step_model.dart';
+import '../pipeline_step_model.dart';
 
 /// Step 4: Extract dates from media files
 ///
