+++ lib/steps/step_05_write_exif/step_05_write_exif.dart
@@ -1,6 +1,6 @@
 import 'package:console_bars/console_bars.dart';
 
-import '../../shared/models/pipeline_step_model.dart';
+import '../pipeline_step_model.dart';
 
 /// Step 5: Write EXIF data to media files
 ///
