+++ lib/steps/step_06_find_albums/step_06_find_albums.dart
@@ -1,4 +1,4 @@
-import '../../shared/models/pipeline_step_model.dart';
+import '../pipeline_step_model.dart';
 
 /// Step 6: Find and merge album relationships
 ///
